Move bulk of decode_aprs main back into the direwolf package
Easier cgo interactions etc.!
